+++ split/split.cc
@@ -9,7 +9,7 @@
 
 //---------------------------------------------------------------------------
 
-size_t TxBuffer::txPacketTypeCounts[SplitHandlerId::COUNT];
+size_t TxBuffer::txPacketTypeCounts[(size_t)SplitHandlerId::COUNT];
 
 //---------------------------------------------------------------------------
 
@@ -19,9 +19,9 @@ bool TxBuffer::Add(SplitHandlerId id, const void *data, size_t length) {
     return false;
   }
 
-  txPacketTypeCounts[id]++;
+  txPacketTypeCounts[(size_t)id]++;
 
-  uint32_t blockHeader = (id << 16) | length;
+  uint32_t blockHeader = ((size_t)id << 16) | length;
   buffer[header.wordCount++] = blockHeader;
 
   memcpy(&buffer[header.wordCount], data, length);
